+++ packages/reactabular-tree/src/index.js
@@ -1,5 +1,6 @@
 export { default as filter } from './filter';
 export { default as getLevel } from './get-level';
+export { default as getChildren } from './get-children';
 export { default as getParents } from './get-parents';
 export { default as hasChildren } from './has-children';
 export { default as pack } from './pack';